[10.x] Generialize paginator classes
